Support big and small immediate constants.
Small constants fall between -2048 and 2047, inclusive.  LIT
instructions with constants in this range will expand to ADDI
instructions in the final instruction stream.  For anything outside this
range, a literal expands into a FWG instruction, which stands for Fetch
Word relative to GP.  This expands to a LD xx,N(GP) instruction in the
final output.  However, since this depends on the GP register being
properly initialized, we need to introduce a new instruction, LDGP,
which will expand into an AUIPC GP,0 instruction.  We don't want to
generate this instruction everywhere though.

One drawback of the approach I've taken is that biglit, will not
coalesce duplicate constants.  So, it'll waste memory.  This is not
expected to be a huge problem in practice.