Tail call elimination.
This pass walks through the list of SCODEs and removes call
CallCode/RfsCode pairs with a single JmpCode.  It's semantically
equivalent, saves code space, and enables control flow without a lot of
special syntax.

BUG: The target addresses used by the JmpXCodes target correct memory
locations BEFORE optimization.  This is because of how I build colon
definitions.  To fix this, I'd need to implement a more dynamic symbol
table and, after collapsing code, I'd need to adjust labels in the
symbol table accordingly.