Draw into our own intermediate buffer.
Don't rely on SDL to retrieve the copy of the video frame buffer for us.
This is dangerous, slow, and frankly, not officially supported at all.
It's better to use the power of wishful thinking, and tell SDL2 how *we*
envision the frame buffer layout to be.  Let SDL2 do the heavy-lifting
of translating RGB surfaces for us, and all that jazz.

I hate having to do all this kind of black magic.  But that's the way
SDL works these days.  Frankly, it's amazing it has any kind of good
performance at all.